OAK-6311 : Move generic part of PropertyStateValue and PropertyValues to oak-store-spi
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1797907 13f79535-47bb-0310-9956-ffa450edef68
