+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/nodetype/NodeTypeIndexLookup.java
@@ -16,7 +16,7 @@
  */
 package org.apache.jackrabbit.oak.plugins.index.nodetype;
 
-import static org.apache.jackrabbit.oak.spi.query.PropertyValues.newName;
+import static org.apache.jackrabbit.oak.plugins.memory.PropertyValues.newName;
 
 import org.apache.jackrabbit.JcrConstants;
 import org.apache.jackrabbit.oak.plugins.index.property.PropertyIndexLookup;
