+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/InImpl.java
@@ -30,7 +30,7 @@ import javax.jcr.PropertyType;
 
 import org.apache.jackrabbit.oak.api.PropertyValue;
 import org.apache.jackrabbit.oak.query.index.FilterImpl;
-import org.apache.jackrabbit.oak.spi.query.PropertyValues;
+import org.apache.jackrabbit.oak.plugins.memory.PropertyValues;
 
 /**
  * A "in" comparison operation.
