+++ oak-core/src/main/java/org/apache/jackrabbit/oak/query/ast/Operator.java
@@ -19,7 +19,7 @@ package org.apache.jackrabbit.oak.query.ast;
 import org.apache.jackrabbit.oak.api.PropertyValue;
 import org.apache.jackrabbit.oak.api.Type;
 import org.apache.jackrabbit.oak.query.fulltext.LikePattern;
-import org.apache.jackrabbit.oak.spi.query.PropertyValues;
+import org.apache.jackrabbit.oak.plugins.memory.PropertyValues;
 
 /**
  * The enumeration of all operators.
