+++ oak-jcr/src/main/java/org/apache/jackrabbit/oak/jcr/query/RowImpl.java
@@ -26,7 +26,7 @@ import javax.jcr.query.Row;
 import org.apache.jackrabbit.oak.api.PropertyValue;
 import org.apache.jackrabbit.oak.api.ResultRow;
 import org.apache.jackrabbit.oak.api.Type;
-import org.apache.jackrabbit.oak.spi.query.PropertyValues;
+import org.apache.jackrabbit.oak.plugins.memory.PropertyValues;
 
 import com.google.common.base.Joiner;
