+++ oak-store-spi/src/test/java/org/apache/jackrabbit/oak/plugins/memory/PropertyStateValueTest.java
@@ -14,10 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.jackrabbit.oak.spi.query;
+package org.apache.jackrabbit.oak.plugins.memory;
 
 import static java.util.Calendar.HOUR_OF_DAY;
-import static org.apache.jackrabbit.oak.spi.query.PropertyValues.newDate;
+import static org.apache.jackrabbit.oak.plugins.memory.PropertyValues.newDate;
 import static org.apache.jackrabbit.util.ISO8601.format;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
